:truck: refactor(openapi): rename from_swagger to from_openapi
The feature targets OpenAPI 3.0 specs, not legacy Swagger 2.0. The
"swagger" name was misleading. Rename the module, public method
(Cli.from_swagger -> Cli.from_openapi), CLI subcommand (from-swagger ->
from-openapi), tests, and docs to match. No back-compat alias since the
feature is unreleased on this branch with no external callers.
