+++ README.md
@@ -89,7 +89,7 @@ No default → positional argument. Has default → `--template` option. Docstri
 - **Config management** — `WithConfig[T]` reads from config files or environment variables (CLI flag > env var > config file > default)
 - **Agent-optimized help** — auto-detects non-TTY output and emits a compact, token-efficient format for LLM agents and scripts
 - **Autogenerated shell completions** — bash, zsh, fish
-- **Swagger/OpenAPI client generation** — `Cli.from_swagger("spec.json")` turns any OpenAPI 3.0 spec into a working CLI; path params become args, query params become options, HTTP calls via stdlib `urllib`
+- **OpenAPI client generation** — `Cli.from_openapi("spec.json")` turns any OpenAPI 3.0 spec into a working CLI; path params become args, query params become options, HTTP calls via stdlib `urllib`
 - **Minimal overhead** — custom parser built from scratch for fast startup; `xclif compile` pre-builds a static manifest to eliminate route-walking cost
 - **Plugin discovery** (planned) — third-party subcommands via entry points (like Git or cargo)
 - **Easy testing** — `command.execute(["greet", "Alice"])` with explicit arg lists, no mocking needed
