+++ docs/index.rst
@@ -29,7 +29,7 @@ file-based routing and a decorator-based API.
    config
    flat-api
    compiler
-   swagger
+   openapi
    mcp
 
 .. toctree::
