+++ src/xclif/__main__.py
@@ -33,21 +33,21 @@ def compile(routes_module: str, output: str = "") -> None:
     print(f"Written: {output_path}")
 
 
-@root.command("from-swagger")
-def from_swagger(spec: str, base_url: str = "", output: str = "") -> None:
-    """Build an xclif CLI from an OpenAPI/Swagger JSON spec.
+@root.command("from-openapi")
+def from_openapi(spec: str, base_url: str = "", output: str = "") -> None:
+    """Build an xclif CLI from an OpenAPI JSON spec.
 
     Parses SPEC (a local JSON file or HTTP(S) URL) and prints the
-    generated command tree.  Use ``Cli.from_swagger()`` in code to use
+    generated command tree.  Use ``Cli.from_openapi()`` in code to use
     the generated CLI programmatically.
     """
-    from xclif.from_swagger import _build_command_tree, load_spec
+    from xclif.from_openapi import _build_command_tree, load_spec
 
     spec_data = load_spec(spec)
     root_cmd = _build_command_tree(spec_data)
 
     if output:
-        _generate_swagger_cli(output, root_cmd, base_url)
+        _generate_openapi_cli(output, root_cmd, base_url)
     else:
         # Print the command tree
         _print_command_tree(root_cmd)
@@ -70,27 +70,27 @@ def _print_command_tree(cmd: Command, indent: int = 0) -> None:
             _print_command_tree(sub, indent + 1)
 
 
-def _generate_swagger_cli(output: str, root_cmd: Command, base_url: str) -> None:
-    """Generate a standalone Python CLI script from the swagger-derived command tree."""
+def _generate_openapi_cli(output: str, root_cmd: Command, base_url: str) -> None:
+    """Generate a standalone Python CLI script from the OpenAPI-derived command tree."""
     # Simple code generation for the static manifest + CLI entry point
     output_path = Path(output)
     root_name = root_cmd.name
 
     lines = [
         "#!/usr/bin/env python3",
-        f'"""CLI generated from Swagger spec — {root_name}."""',
+        f'"""CLI generated from OpenAPI spec — {root_name}."""',
         "",
         "from pathlib import Path",
         "",
         "from xclif import Cli, command",
         "from xclif.command import Command",
-        "from xclif.from_swagger import _build_command_tree, load_spec",
+        "from xclif.from_openapi import _build_command_tree, load_spec",
         "",
         f'SPEC_PATH = Path(__file__).with_suffix(".json")',
         "",
         "def main() -> None:",
         "    if SPEC_PATH.is_file():",
-        "        cli = Cli.from_swagger(str(SPEC_PATH)",
+        "        cli = Cli.from_openapi(str(SPEC_PATH)",
     ]
 
     if base_url:
